improve the handling method that currency switching
