+++ packages/web/src/components/molecules/WithdrawForm/index.spec.tsx
@@ -6,6 +6,7 @@ import 'src/__mocks__/window/matchMedia.mock'
 
 jest.mock('src/fixtures/dev-kit/hooks')
 jest.mock('src/fixtures/uniswap/hooks')
+jest.mock('src/fixtures/currency/hooks')
 
 describe(`${WithdrawForm.name}`, () => {
   test('Snapshot', () => {
