+++ packages/web/src/components/organisms/PropertyCardList/index.spec.tsx
@@ -5,6 +5,7 @@ import 'src/__mocks__/window/matchMedia.mock'
 
 jest.mock('src/fixtures/dev-kit/hooks')
 jest.mock('@dev/graphql')
+jest.mock('src/fixtures/currency/hooks')
 
 describe(`${PropertyCardList.name}`, () => {
   test.skip('Snapshot', () => {
