+++ packages/web/src/fixtures/currency/__mocks__/hooks.ts
@@ -1,0 +1,3 @@
+export const useCurrency = () => {
+  return { currency: 'DEV', toCurrency: (x: any) => x }
+}
